refactor: register bundles from config array instead of using DynamicProcessor
Change the logic to register bundles from the config array of strings received from the engine, rather than reading them through the DynamicProcessor. This adjustment is part of separating the package. In the future, the logic will be updated to allow packages to define their own processors and bundles within package.json.